pyproject.toml
[tool.poetry]
name = "pat-project"
version = "0.1.0"
description = ""
authors = ["baha_2002 <baktiarzaksylykov22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
django = "^5.0.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
